requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
opencv-python>=4.7.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Bar
tqdm>=4.65.0

# Configuration
PyYAML>=6.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.22.0

# Testing
pytest>=7.3.0

# Utilities
scikit-learn>=1.2.0

# Grad-CAM
grad-cam>=1.4.0
